refactor: 미션 제약에 따른 패키지 구조 수정
`서비스와 데이터의 특성을 고려했을 때 쿠폰과 회원에게 발급된 쿠폰의 양이 다를 것으로 예상되어, 두 데이터의 저장소를 분리하는 작업을 계획 중이다.`
