+++ src/main/java/coupon/coupon/domain/Category.java
@@ -1,0 +1,9 @@
+package coupon.coupon.domain;
+
+public enum Category {
+
+    FASHION,
+    ELECTRONICS,
+    FURNITURE,
+    FOOD,
+}
