+++ src/main/java/coupon/coupon/domain/DiscountPolicy.java
@@ -1,0 +1,7 @@
+package coupon.coupon.domain;
+
+public interface DiscountPolicy {
+
+    void validatePolicy(long fixedDiscountAmount, long minimumOrderPrice)
+            throws DiscountPolicyViolationException;
+}
