+++ src/main/java/coupon/coupon/domain/DiscountPolicyViolationException.java
@@ -1,0 +1,8 @@
+package coupon.coupon.domain;
+
+public class DiscountPolicyViolationException extends RuntimeException {
+
+    public DiscountPolicyViolationException(String message) {
+        super(message);
+    }
+}
