+++ src/main/java/coupon/coupon/repository/IssuedCouponRepository.java
@@ -1,0 +1,8 @@
+package coupon.coupon.repository;
+
+import org.springframework.data.repository.Repository;
+
+public interface IssuedCouponRepository extends Repository<IssuedCouponEntity, Long> {
+
+    IssuedCouponEntity save(IssuedCouponEntity issuedCoupon);
+}
