+++ src/main/java/coupon/coupon/service/CouponCreationRequest.java
@@ -1,0 +1,14 @@
+package coupon.coupon.service;
+
+import coupon.coupon.domain.Category;
+import java.time.LocalDate;
+
+public record CouponCreationRequest(
+        String name,
+        Category category,
+        long discountPrice,
+        long minimumOrderPrice,
+        LocalDate issuableDate,
+        LocalDate expirationDate
+) {
+}
