--- src/main/java/coupon/domain/Category.java
@@ -1,9 +1,0 @@
-package coupon.domain;
-
-public enum Category {
-
-    FASHION,
-    ELECTRONICS,
-    FURNITURE,
-    FOOD,
-}
