--- src/main/java/coupon/domain/coupon/DiscountPolicy.java
@@ -1,7 +1,0 @@
-package coupon.domain.coupon;
-
-public interface DiscountPolicy {
-
-    void validatePolicy(long fixedDiscountAmount, long minimumOrderPrice)
-            throws DiscountPolicyViolationException;
-}
