--- src/main/java/coupon/domain/coupon/DiscountPolicyViolationException.java
@@ -1,8 +1,0 @@
-package coupon.domain.coupon;
-
-public class DiscountPolicyViolationException extends RuntimeException {
-
-    public DiscountPolicyViolationException(String message) {
-        super(message);
-    }
-}
