--- src/main/java/coupon/repository/IssuedCouponRepository.java
@@ -1,8 +1,0 @@
-package coupon.repository;
-
-import org.springframework.data.repository.Repository;
-
-public interface IssuedCouponRepository extends Repository<IssuedCouponEntity, Long> {
-
-    IssuedCouponEntity save(IssuedCouponEntity issuedCoupon);
-}
